Add gravatar to Person model
Person now have the field email that is used to get the gravatar url of the person.
Also, the name of the person is generated from email address.

Adding the rake task `rails update_people_email` that set email to people in DB.
This task update also the person name.